moving component around
