+++ src/sections/SelectStage/types.ts
